Cleaning up tests.
