+++ tests/System.Text.Primitives.Tests/Bugs.cs
